fix: PR review comment fixes, styling
