+++ src/components/ConnectProvider.svelte
@@ -6,8 +6,8 @@
 </script>
 
 <Modal id="connect-provider" {close} {isOpen}>
-  <div class="flex flex-col gap-7">
-    <h2 class="section-title-underlined">Connect Provider</h2>
+  <span slot="title">Connect Provider</span>
+  <svelte:fragment slot="body">
     <!-- Select a network dropdown goes here -->
     <!-- Should default to current network -->
     <DropDownMenu label="Select a Network" options={['option1']} />
@@ -20,5 +20,5 @@
       <button class="btn-primary" on:click={close}>Connect</button>
       <button class="btn-no-fill" on:click={close}>Cancel</button>
     </div>
-  </div>
+  </svelte:fragment>
 </Modal>
