+++ src/style/app.css
@@ -90,12 +90,13 @@
 
   input,
   [type='text'],
+  [type='number'],
   select {
-    @apply h-[30px] w-[320px] overflow-ellipsis rounded-md border-none py-0 pl-2 pr-7 text-sm text-black outline-none disabled:text-disabled disabled:opacity-85;
+    @apply h-[30px] w-[320px] overflow-ellipsis rounded-md border-none py-0 pl-2 text-sm text-black outline-none disabled:text-disabled disabled:opacity-85;
   }
 
   select {
-    @apply cursor-pointer;
+    @apply cursor-pointer pr-7;
   }
 
   ::placeholder {
